requirements.txt
NUMPY
python-jobspy
openai
pandas
python-dotenv
markdown2
langchain
langchain_openai
pydantic
bs4
fake_useragent
